Refactor user search queries out of user search method
